update gitignore file; add license info
