+++ tenant/angular4/src/app/repository/create/create.component.html
@@ -1,3 +1,18 @@
+
+<!--lCopyright 2016 - 2017 Huawei Technologies Co., Ltd. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+imitations under the License.-->
+
 <div class="d3-design" #design></div>
 <md-card>
   <md-card-title>Create Repository</md-card-title>
