--- tenant/angular4/src/app/repository/create/repo.service.ts
@@ -1,7 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class RepoService {
-
-      addRepo(): void {} // stub
-}
