--- tenant/angular4/src/app/repository/create/repo.ts
@@ -1,7 +1,0 @@
-export class Repo {
-  name: string;
-  repo_type: number;
-  islimit:boolean;
-  Description:string
-  ishavekey:boolean;
-}
